3PH: Implement forwarding.
When we receive a third-party cap, we don't accept it right away. We lazily wait until it is actually called, then accept it. If we end up forwarding the cap on to yet another party before accepting it, then we check if the VatNetwork supports forwarding without forming a connection back to the original host. If so, we use that.

This allows a capability to be passed through many hands before actually being invoked, without forcing each intermediate node to make a connection all the way back to the host.