Refactor: Simplify code assuming promises don't resolve to resolved promises.
As described in the comments in rpc.capnp, you cannot re-export a promise when that promise was already resolved in the past, because this would likely confuse the other end.

This has always been true, but not called out. Luckily, it has also always been true that the (C++) implementation doesn't do this. The sending side always collapses all already-resolved promises on its end.

Meanwhile, RpcPromise::resolve() contained a bunch of code to handle the case of resolving to a promise that's already resolved. This situation was never possible, due to the above de facto (now de jure) rule, so we can simplify it.