3PH: Explicitly store vine info on export table.
In previous commits I was using a horrible hack where I'd `.attach()` stuff to the vine's ClientHook in order to keep that stuff alive until the vine was dropped. This felt bad because ClientHooks are refcounted but the addRef() would lose the attachment -- however, that was actually a feature and not a bug, because we really only wanted to keep the attachment on the export table, not on any other references cloned from there.

A bigger problem comes up with pipelining. I had made it so that when a 3PH was forwarded, the export for the vine actually pointed to the DeferredThirdPartyClient object. But this meant that a call sent to the vine would actually cause the DeferredThirdPartyClient to trigger an Accept, instead of passing the call along the vine as expected.

The problem: Consider a long promise chain through a bunch of nodes. Pipelined calls are being made along the chain. Then the end of the chain resolves to a third-party handoff. That handoff propagates back through the chain. We want to take advantage of forwarding here; we don't want every node in the chain to form a connection to the capability host. However, because there are pipelined calls travelling in the opposite direction at the same time, one of them might collide with the resolution cascade. At that point it'd hit a DeferredThirdPartyClient on whatever middleman node it's on, and that object would say "whoa, I got a call, I'd better actually accept the capability", and would form a spurious connection.

A second problem comes with embargoes: A three-party embargo actually needs to follow the vine back to find the original `Provide` operation that started the three-party handoff, and then needs to follow that `Provide`. But if the disembargo hit a `DeferredThirdPartyClient` object, and that object had in the meantime decided to perform an accept and so was no longer hanging onto the original vine, then the disembargo could not continue following the vine. Also, when it gets to the end of the vine, it needs to know about the Provide operation's `QuestionRef`, which was previously hidden in an attachment on the `Own<ClientHook>`, so couldn't have been accessed. Now it is stored separately on the export table.